added enabled variable to disable the bar by default
