--- server/linux_x11/plugins/recognition_result.yapsy-plugin
@@ -1,8 +1,0 @@
-[Core]
-Name = recognition_result
-Module = recognitionresult
-
-[Documentation]
-Author = Facundo Quiroga
-Version = 0.1
-Description = Displays the recognized phrases of the client in the server. Tested only in Ubuntu 18.10. Requires python-tk or equivalent.
