+++ server/linux_x11/plugins/recognitionbar/recognitionbar_tk.py
